requirements.txt
Pillow
numpy
tensorflow
streamlit
PyQt5 
PyQtWebEngine 